Make logsumexp and logaddexp stable by construction
Both built a bare log(sum(exp(x))) and left the max subtraction to
local_log_sum_exp, which runs after grad. The gradient was therefore taken from
the raw form as exp(x) / sum(exp(x)), which overflows past log(float64 max) and
is 0/0 for very negative x. #2288 only addressed the forward value.

LogSumExp and LogAddExp subtract the maximum as built, so their gradient is
stable without a pullback override. The rewrites that emit them run at stabilize
only; at specialize they would re-wrap what late_inline_OpFromGraph just
inlined. Both helpers move next to the Ops in tensor/special.py, still
importable as pt.logsumexp and pt.logaddexp.
