+++ pytensor/tensor/rewriting/special.py
@@ -1,9 +1,9 @@
 from pytensor.graph.rewriting.basic import copy_stack_trace, node_rewriter
 from pytensor.scalar.basic import Exp
 from pytensor.tensor.elemwise import DimShuffle, Elemwise
-from pytensor.tensor.math import Sum, log, true_div
+from pytensor.tensor.math import Sum, add, exp, log, true_div
 from pytensor.tensor.rewriting.basic import register_stabilize
-from pytensor.tensor.special import Softmax, log_softmax
+from pytensor.tensor.special import Softmax, log_softmax, logaddexp, logsumexp
 from pytensor.tensor.subtensor import (
     AdvancedSubtensor,
     Subtensor,
@@ -94,4 +94,52 @@ def local_softmax_stabilize(fgraph, node):
 
     ret = Softmax(axis=axis)(x)
     copy_stack_trace(node.outputs, ret)
+    return [ret]
+
+
+@register_stabilize("symbolic_op_recognition", "fast_compile")
+@node_rewriter([log])
+def local_log_add_exp(fgraph, node):
+    """``log(exp(x) + exp(y) + exp(z)) -> logaddexp(x, y, z)``.
+
+    TODO: in canonicalize, change log10 and log2 -> log
+    """
+    z = node.inputs[0]
+    if z.owner and z.owner.op == add:
+        zi = z.owner.inputs
+        pre_exp = [x.owner.inputs[0] for x in zi if x.owner and x.owner.op == exp]
+        # all arguments to add are exp(<something>)
+        if len(pre_exp) == len(zi):
+            return [logaddexp(*pre_exp)]
+
+
+@register_stabilize("symbolic_op_recognition", "fast_compile")
+@node_rewriter([log])
+def local_log_sum_exp(fgraph, node):
+    """``log(sum_i(exp(x_i))) -> logsumexp(x)``."""
+    sum_node = node.inputs[0].owner
+    # If the sum has keepdims=True, there might be a dimshuffle
+    if sum_node and isinstance(sum_node.op, DimShuffle):
+        dimshuffle_op = sum_node.op
+        sum_node = sum_node.inputs[0].owner
+    else:
+        dimshuffle_op = None
+
+    if not (sum_node and isinstance(sum_node.op, Sum)):
+        return
+
+    exp_node, axis = sum_node.inputs[0].owner, sum_node.op.axis
+    if not (
+        exp_node
+        and isinstance(exp_node.op, Elemwise)
+        and isinstance(exp_node.op.scalar_op, Exp)
+    ):
+        return
+
+    ret = logsumexp(exp_node.inputs[0], axis=axis)
+
+    # Restore the dimshuffle op, if any.
+    if dimshuffle_op:
+        ret = dimshuffle_op(ret)
+
     return [ret]
