+++ tests/tensor/rewriting/test_special.py
@@ -1,8 +1,11 @@
+import copy
+
 import numpy as np
 import pytest
 import scipy.special
 
 import pytensor
+import pytensor.scalar as ps
 from pytensor import shared
 from pytensor.compile import optdb
 from pytensor.compile.mode import get_mode
@@ -10,9 +13,11 @@ from pytensor.configdefaults import config
 from pytensor.graph.fg import FunctionGraph
 from pytensor.graph.rewriting.basic import check_stack_trace
 from pytensor.graph.rewriting.db import RewriteDatabaseQuery
-from pytensor.tensor.math import exp, log
+from pytensor.tensor.elemwise import DimShuffle
+from pytensor.tensor.math import Max, exp, log
+from pytensor.tensor.math import sum as pt_sum
 from pytensor.tensor.special import LogSoftmax, Softmax, softmax
-from pytensor.tensor.type import matrix
+from pytensor.tensor.type import TensorType, dvector, matrix, tensor3, vector
 from tests import unittest_tools as utt
 
 
@@ -114,3 +119,128 @@ def test_softmax_graph():
         return pytensor.grad(None, x, known_grads={y: inputs})
 
     utt.verify_grad(f, [rng.random((3, 4))])
+
+
+@pytest.mark.parametrize("mode", ["FAST_COMPILE", "FAST_RUN"])
+def test_local_log_add_exp(mode):
+    m = get_mode(mode).excluding("fusion")
+    m = copy.copy(m)
+    # No need to put them back as we have a new object
+    m.check_isfinite = False
+
+    # check some basic cases
+    x = dvector()
+    y = dvector()
+    f = pytensor.function([x, y], log(exp(x) + exp(y)), mode=m)
+
+    # test that it gives the correct result when it doesn't overflow
+    f([10], [10])  # doesn't causes overflow
+    utt.assert_allclose(f([10], [10]), 10 + np.log1p(1))
+
+    assert np.isfinite(f([10000], [10000]))  # causes overflow if handled incorrectly
+    utt.assert_allclose(f([10000], [10000]), 10000 + np.log1p(1))
+
+    # test that when max = +-inf, rewritten output still works correctly
+    assert f([-np.inf], [-np.inf]) == -np.inf
+    assert f([np.inf], [np.inf]) == np.inf
+    assert f([np.inf], [-np.inf]) == np.inf
+
+    # test that it also works with more than two args
+    x = dvector()
+    y = dvector()
+    f = pytensor.function(
+        [x, y], log(exp(x) + exp(y) + exp(x - y) + exp(x + y)), mode=m
+    )
+
+    assert np.isfinite(f([10000], [10000]))  # causes overflow if handled incorrectly
+    utt.assert_allclose(f([10000], [10000]), 20000)
+
+    # TODO: test that the rewrite works in the presence of broadcasting.
+
+
+def compile_graph_log_sum_exp(x, axis, dimshuffle_op=None, mode="FAST_RUN"):
+    sum_exp = pt_sum(exp(x), axis=axis)
+    if dimshuffle_op:
+        sum_exp = dimshuffle_op(sum_exp)
+    y = log(sum_exp)
+    return pytensor.function([x], y, mode=mode)
+
+
+def check_max_log_sum_exp(x, axis, dimshuffle_op=None):
+    f = compile_graph_log_sum_exp(x, axis, dimshuffle_op)
+
+    fgraph = f.maker.fgraph.toposort()
+    for node in fgraph:
+        if hasattr(node.op, "scalar_op") and node.op.scalar_op == ps.basic.maximum:
+            return
+
+        if isinstance(node.op, Max):
+            return
+
+    # TODO FIXME: Refactor this test so that it makes a direct assertion and
+    # nothing more.
+    raise AssertionError("No maximum detected after log_sum_exp rewrite")
+
+
+def test_local_log_sum_exp_maximum():
+    """Test that the rewrite is applied by checking the presence of the maximum."""
+    x = tensor3("x")
+    check_max_log_sum_exp(x, axis=(0,), dimshuffle_op=None)
+    check_max_log_sum_exp(x, axis=(1,), dimshuffle_op=None)
+    check_max_log_sum_exp(x, axis=(2,), dimshuffle_op=None)
+    check_max_log_sum_exp(x, axis=(0, 1), dimshuffle_op=None)
+    check_max_log_sum_exp(x, axis=(0, 1, 2), dimshuffle_op=None)
+
+    # If a transpose is applied to the sum
+    transpose_op = DimShuffle(input_ndim=2, new_order=(1, 0))
+    check_max_log_sum_exp(x, axis=2, dimshuffle_op=transpose_op)
+
+    # If the sum is performed with keepdims=True
+    x = TensorType(dtype="floatX", shape=(None, 1, None))("x")
+    sum_keepdims_op = x.sum(axis=(0, 1), keepdims=True).owner.op
+    check_max_log_sum_exp(x, axis=(0, 1), dimshuffle_op=sum_keepdims_op)
+
+
+def test_local_log_sum_exp_near_one():
+    """Test that the rewritten result is correct around 1.0."""
+
+    x = tensor3("x")
+    x_val = 1.0 + np.random.random((4, 3, 2)).astype(config.floatX) / 10.0
+
+    f = compile_graph_log_sum_exp(x, axis=(1,))
+    naive_ret = np.log(np.sum(np.exp(x_val), axis=1))
+    rewritten_ret = f(x_val)
+    assert np.allclose(naive_ret, rewritten_ret)
+
+    # If a transpose is applied
+    transpose_op = DimShuffle(input_ndim=2, new_order=(1, 0))
+    f = compile_graph_log_sum_exp(x, axis=(1,), dimshuffle_op=transpose_op)
+    naive_ret = np.log(np.sum(np.exp(x_val), axis=1).T)
+    rewritten_ret = f(x_val)
+    assert np.allclose(naive_ret, rewritten_ret)
+
+
+@pytest.mark.parametrize("mode", ["FAST_COMPILE", "FAST_RUN"])
+@pytest.mark.parametrize(
+    "x_val", ([-800.0, 800.0], [-800.0, -805.0]), ids=["overflow", "underflow"]
+)
+def test_local_log_sum_exp_large(x_val, mode):
+    """Test that the rewrite result is correct for values the naive graph can't represent."""
+    x = vector("x")
+    f = compile_graph_log_sum_exp(x, axis=0, mode=mode)
+
+    x_val = np.array(x_val, dtype=config.floatX)
+
+    rewritten_ret = f(x_val)
+    np.testing.assert_allclose(rewritten_ret, scipy.special.logsumexp(x_val), rtol=1e-5)
+
+
+@pytest.mark.parametrize("mode", ["FAST_COMPILE", "FAST_RUN"])
+def test_local_log_sum_exp_inf(mode):
+    """Test that when max = +-inf, the rewritten output still works correctly."""
+    x = vector("x")
+    f = compile_graph_log_sum_exp(x, axis=0, mode=mode)
+
+    assert f([-np.inf, -np.inf]) == -np.inf
+    assert f([np.inf, np.inf]) == np.inf
+    assert f([-np.inf, np.inf]) == np.inf
